git-svn-id: https://dotnetnuke.svn.codeplex.com/svn@68011 142e9c35-602f-4880-9f37-ec48a24daa6a

+++ Trunk/Community/Library/Services/Log/EventLog/LogController.cs
@@ -292,7 +292,7 @@ namespace DotNetNuke.Services.Log.EventLog
                                                             logTypeConfig.EmailNotificationIsActive,
                                                             Convert.ToString(logTypeConfig.NotificationThreshold),
                                                             Convert.ToString(logTypeConfig.NotificationThresholdTime),
-                                                            Convert.ToString(logTypeConfig.NotificationThresholdTimeType),
+                                                            Convert.ToString((int)logTypeConfig.NotificationThresholdTimeType),
                                                             logTypeConfig.MailFromAddress,
                                                             logTypeConfig.MailToAddress);
         }
@@ -373,7 +373,7 @@ namespace DotNetNuke.Services.Log.EventLog
                                                                logTypeConfig.EmailNotificationIsActive,
                                                                Convert.ToString(logTypeConfig.NotificationThreshold),
                                                                Convert.ToString(logTypeConfig.NotificationThresholdTime),
-                                                               Convert.ToString(logTypeConfig.NotificationThresholdTimeType),
+                                                               Convert.ToString((int)logTypeConfig.NotificationThresholdTimeType),
                                                                logTypeConfig.MailFromAddress,
                                                                logTypeConfig.MailToAddress);
         }
